improven JBON format using the marker byte to include the size of small string,arrays or objects

+++ src/main/java/net/woggioni/worth/serialization/binary/BinaryMarker.java
@@ -15,8 +15,8 @@ public enum BinaryMarker {
     EmptyArray(0xaf),
     LargeArray(0xff);
 
-    public byte value;
+    public int value;
     BinaryMarker(int b) {
-        value = (byte) b;
+        value = b;
     }
 }
